Updated Pages folder and UIclasses (#30)

+++ IQDHackathon/IQDHackathon/UiClasses/clsTitle.cs
@@ -14,8 +14,7 @@
             Score = score;
             NumberOfAnswer = numberOfAnswer;
             QuestionTitle = questionTitle;
-            ScoreForBranchOrPint = (byte)(Score / NumberOfAnswer);
+            ScoreForBranchOrPint = (float)(Score / NumberOfAnswer);
         }
-
     }
 }
